+++ src/main/kotlin/cloud/glitchdev/rfu/achievement/AchievementCategory.kt
@@ -3,5 +3,6 @@ package cloud.glitchdev.rfu.achievement
 enum class AchievementCategory {
     GENERAL,
     ISLE,
+    HOT_SPOT,
     SPECIAL
 }